# xorBlock does not run on 386 arch