The xorBlock function expects the wordSize to be 8 bytes. On arch 386
this size is 4 bytes and xorBlock will panic. xorWords is copied from
golang's crypto/cipher/xor.go.