@jacobsa 2016-02-17
Thanks for the pull request. This appears to undo some of the performance improvements of 61fbfe16bbddd8aa7f925049a82dd3955b07e2a6, which intentionally eliminated the use of slices (since we don't want to pay for bounds checking when we statically know the length).

The benchmarks are a bit noisy, but they show the effect. Before:

```
% go test ./siv -test.run asdf -test.benchmem -test.bench 16MiB -test.benchtime 5s
PASS
BenchmarkDecrypt16MiB-8      100          82948416 ns/op         202.26 MB/s    33556240 B/op         29 allocs/op
BenchmarkEncrypt16MiB-8      100          79892663 ns/op         210.00 MB/s    20981520 B/op         28 allocs/op
```

After:

```
BenchmarkDecrypt16MiB-8      100          87123995 ns/op         192.57 MB/s    33556256 B/op         30 allocs/op
BenchmarkEncrypt16MiB-8      100          86094686 ns/op         194.87 MB/s    20981520 B/op         28 allocs/op
```

Could you instead rework this so that the existing `xorBlock` function is in hash_amd64.go, and put a new version with the same signature in hash_386.go? We could handle other architectures as needed.