@itiserik 2016-02-17
Before (using xorWords):
```
go test ./siv -test.run asdf -test.benchmem -test.bench 16MiB -test.benchtime 5s
PASS
BenchmarkDecrypt16MiB-4        2        3928990273 ns/op           4.27 MB/s    33556112 B/op         29 allocs/op
BenchmarkEncrypt16MiB-4        2        3910084514 ns/op           4.29 MB/s    20981392 B/op         28 allocs/op
```
With this patch applied:
```
go test ./siv -test.run asdf -test.benchmem -test.bench 16MiB -test.benchtime 5s
PASS
BenchmarkDecrypt16MiB-4        2        3813656927 ns/op           4.40 MB/s    33556096 B/op         28 allocs/op
BenchmarkEncrypt16MiB-4        2        3761727765 ns/op           4.46 MB/s    20981408 B/op         29 allocs/op

```
Your milage may vary (tm)